refactor(DataStore): Split concerns into files
Closes #21
