+++ libs/data-store/src/lib/utils.ts
@@ -1,0 +1,3 @@
+export function deepCopy<Type>(obj: Type): Type {
+  return JSON.parse(JSON.stringify(obj));
+}
